WebAssembly: Hack: disable all atomic instructions
This matches what Clang does.

This doesn't check that the target is Wasm; will need to do that.

Patcheng's port seems to have working? atomics, so it's probably possible
to get atomics working, but this is fine for now.

This still doesn't get the stdlib to compile, but at least the error now matches
what I get when I run clang on the -emit-ir LLVM IR output.